Move OsfWebRenderer to framework.routing
