+++ framework/routing/__init__.py
@@ -619,3 +619,84 @@ class WebRenderer(Renderer):
         data.update({key: val for key, val in extra_data.iteritems() if key not in data})
 
         return self._render(data, template_name)
+
+def get_globals():
+    """Context variables that are available for every template rendered by
+    OSFWebRenderer.
+    """
+    user = _get_current_user()
+    user_institutions = [{'id': inst._id, 'name': inst.name, 'logo_path': inst.logo_path_rounded_corners} for inst in user.affiliated_institutions.all()] if user else []
+    location = geolite2.lookup(request.remote_addr) if request.remote_addr else None
+    if request.host_url != settings.DOMAIN:
+        try:
+            inst_id = (Institution.find_one(Q('domains', 'eq', request.host.lower())))._id
+            request_login_url = '{}institutions/{}'.format(settings.DOMAIN, inst_id)
+        except NoResultsFound:
+            request_login_url = request.url.replace(request.host_url, settings.DOMAIN)
+    else:
+        request_login_url = request.url
+    return {
+        'private_link_anonymous': is_private_link_anonymous_view(),
+        'user_name': user.username if user else '',
+        'user_full_name': user.fullname if user else '',
+        'user_id': user._id if user else '',
+        'user_locale': user.locale if user and user.locale else '',
+        'user_timezone': user.timezone if user and user.timezone else '',
+        'user_url': user.url if user else '',
+        'user_gravatar': profile_views.current_user_gravatar(size=25)['gravatar_url'] if user else '',
+        'user_email_verifications': user.unconfirmed_email_info if user else [],
+        'user_api_url': user.api_url if user else '',
+        'user_entry_point': metrics.get_entry_point(user) if user else '',
+        'user_institutions': user_institutions if user else None,
+        'display_name': get_display_name(user.fullname) if user else '',
+        'anon': {
+            'continent': getattr(location, 'continent', None),
+            'country': getattr(location, 'country', None),
+        },
+        'use_cdn': settings.USE_CDN_FOR_CLIENT_LIBS,
+        'sentry_dsn_js': settings.SENTRY_DSN_JS if sentry.enabled else None,
+        'dev_mode': settings.DEV_MODE,
+        'allow_login': settings.ALLOW_LOGIN,
+        'cookie_name': settings.COOKIE_NAME,
+        'status': status.pop_status_messages(),
+        'prev_status': status.pop_previous_status_messages(),
+        'domain': settings.DOMAIN,
+        'api_domain': settings.API_DOMAIN,
+        'disk_saving_mode': settings.DISK_SAVING_MODE,
+        'language': language,
+        'noteworthy_links_node': settings.NEW_AND_NOTEWORTHY_LINKS_NODE,
+        'popular_links_node': settings.POPULAR_LINKS_NODE,
+        'web_url_for': util.web_url_for,
+        'api_url_for': util.api_url_for,
+        'api_v2_url': util.api_v2_url,  # URL function for templates
+        'api_v2_base': util.api_v2_url(''),  # Base url used by JS api helper
+        'sanitize': sanitize,
+        'sjson': lambda s: sanitize.safe_json(s),
+        'webpack_asset': paths.webpack_asset,
+        'waterbutler_url': settings.WATERBUTLER_URL,
+        'login_url': cas.get_login_url(request_login_url),
+        'reauth_url': util.web_url_for('auth_logout', redirect_url=request.url, reauth=True),
+        'profile_url': cas.get_profile_url(),
+        'enable_institutions': settings.ENABLE_INSTITUTIONS,
+        'keen': {
+            'public': {
+                'project_id': settings.KEEN['public']['project_id'],
+                'write_key': settings.KEEN['public']['write_key'],
+            },
+            'private': {
+                'project_id': settings.KEEN['private']['project_id'],
+                'write_key': settings.KEEN['private']['write_key'],
+            },
+        },
+        'maintenance': maintenance.get_maintenance(),
+        'recaptcha_site_key': settings.RECAPTCHA_SITE_KEY
+    }
+
+class OsfWebRenderer(WebRenderer):
+    """Render a Mako template with OSF context vars.
+
+    :param trust: Optional. If ``False``, markup-safe escaping will be enabled
+    """
+    def __init__(self, *args, **kwargs):
+        kwargs['data'] = get_globals
+        super(OsfWebRenderer, self).__init__(*args, **kwargs)
