+++ tests/framework_tests/test_modular_templates.py
@@ -4,23 +4,17 @@
 These require a test db because they use Session objects.
 '''
 import json
-import unittest
 import os
+import unittest
 
 import flask
-from lxml.html import fragment_fromstring
 import werkzeug.wrappers
-
 from framework.exceptions import HTTPError, http
-from framework.routing import (
-    Renderer, JSONRenderer, WebRenderer,
-    render_mako_string,
-)
-
-from tests.base import AppTestCase, OsfTestCase
-
 # Need to use OsfWebRenderer to get global variables
-from website.routes import OsfWebRenderer
+from framework.routing import (JSONRenderer, OsfWebRenderer, Renderer,
+                               WebRenderer, render_mako_string)
+from lxml.html import fragment_fromstring
+from tests.base import AppTestCase, OsfTestCase
 
 HERE = os.path.dirname(os.path.abspath(__file__))
 TEMPLATES_PATH = os.path.join(HERE, 'templates')
