+++ website/addons/forward/routes.py
@@ -1,8 +1,6 @@
 # -*- coding: utf-8 -*-
 """Forward addon routes."""
-from framework.routing import Rule, json_renderer
-from website.routes import OsfWebRenderer
-
+from framework.routing import OsfWebRenderer, Rule, json_renderer
 from website.addons.forward import views
 
 api_routes = {
