+++ website/addons/wiki/routes.py
@@ -4,8 +4,7 @@ Routes associated with the wiki page
 
 import os
 
-from framework.routing import Rule, json_renderer
-from website.routes import OsfWebRenderer
+from framework.routing import OsfWebRenderer, Rule, json_renderer
 
 from . import views
